+++ lib/screen.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+mixin Screen {
+  void showToast(String message, GlobalKey<ScaffoldState> _scaffoldKey) {
+    _scaffoldKey.currentState.showSnackBar(
+      SnackBar(
+        content: Text(message),
+      ),
+    );
+  }
+}